Deleted Toolbox class.
