+++ docs/Reverse-Records.md
@@ -6,9 +6,9 @@ The DNS Library can natively handle reverse IP records for both IPv4 and IPv6.
 ```php
 use Badcow\DNS\Rdata\Factory;
 use Badcow\DNS\{Zone, ResourceRecord, AlignedBuilder, Classes, ZoneBuilder};
-use Badcow\DNS\Ip\Toolbox;
+use Badcow\DNS\Ip\PTR;
 
-$parent = Toolbox::reverseIpv4('158.133.7'); //Returns 7.133.158.in-addr.arpa.
+$parent = PTR::reverseIpv4('158.133.7'); //Returns 7.133.158.in-addr.arpa.
 
 $zone = new Zone($parent, 10800);
 
@@ -43,20 +43,20 @@ $TTL 10800
 ```php
 use Badcow\DNS\Rdata\Factory;
 use Badcow\DNS\{Zone, ResourceRecord, AlignedBuilder, Classes, ZoneBuilder};
-use Badcow\DNS\Ip\Toolbox;
+use Badcow\DNS\Ip\PTR;
 
-//$parent = Toolbox::reverseIpv4('158.133.7'); //Returns 7.133.158.in-addr.arpa.
-$parent = Toolbox::reverseIpv6('2001:acad:5889:0:0:0:0');
+//$parent = PTR::reverseIpv4('158.133.7'); //Returns 7.133.158.in-addr.arpa.
+$parent = PTR::reverseIpv6('2001:acad:5889:0:0:0:0');
 
 $zone = new Zone($parent, 10800);
 
 $resourceRecords = [
-    new ResourceRecord(Toolbox::reverseIpv6('1', false), Factory::Ptr('gw-01.badcow.co.')),
-    new ResourceRecord(Toolbox::reverseIpv6('2', false), Factory::Ptr('gw-02.badcow.co.')),
-    new ResourceRecord(Toolbox::reverseIpv6('bad', false), Factory::Ptr('badcow.co.')),
-    new ResourceRecord(Toolbox::reverseIpv6('ff', false), Factory::Ptr('mail.badcow.co.'), 3600, Classes::INTERNET),
-    new ResourceRecord(Toolbox::reverseIpv6('aa1', false), Factory::Ptr('esw-01.badcow.co.')),
-    new ResourceRecord(Toolbox::reverseIpv6('aa2', false), Factory::Ptr('esw-02.badcow.co.')),
+    new ResourceRecord(PTR::reverseIpv6('1', false), Factory::Ptr('gw-01.badcow.co.')),
+    new ResourceRecord(PTR::reverseIpv6('2', false), Factory::Ptr('gw-02.badcow.co.')),
+    new ResourceRecord(PTR::reverseIpv6('bad', false), Factory::Ptr('badcow.co.')),
+    new ResourceRecord(PTR::reverseIpv6('ff', false), Factory::Ptr('mail.badcow.co.'), 3600, Classes::INTERNET),
+    new ResourceRecord(PTR::reverseIpv6('aa1', false), Factory::Ptr('esw-01.badcow.co.')),
+    new ResourceRecord(PTR::reverseIpv6('aa2', false), Factory::Ptr('esw-02.badcow.co.')),
 ];
 
 $zone->fromArray($resourceRecords);
